remove ActivityContext and replace with SessionContext
