+++ src/components/ActivityBody/ActivityBody.jsx
@@ -1,13 +1,13 @@
 import React, { useContext } from 'react';
-import { ActivityContext } from '../../contexts/ActivityContext';
+import { SessionContext } from '../../contexts/SessionContext';
 import ActivityLinks from '../ActivityLinks/ActivityLinks';
 import ActivityMain from '../ActivityMain/ActivityMain';
 import ConfirmationForm from '../ConfirmationForm/ConfirmationForm';
 
 const ActivityBody = () => {
-  const { started, finished } = useContext(ActivityContext);
+  const { started, finished } = useContext(SessionContext);
   
-  return (
+  return ( 
     <>
       {started === true && <ActivityMain />}
       {started === false && finished === null && <ConfirmationForm />}
