+++ src/components/ActivityMain/ActivityMain.jsx
@@ -1,14 +1,14 @@
 import React, { useContext, useEffect } from 'react';
-import { ActivityContext } from '../../contexts/ActivityContext';
 import { SessionContext } from '../../contexts/SessionContext';
+
 import * as U from '../../components/TuxComponents/UniversalComponents';
 import Question from '../../components/Question/Question';
 import InjectHTML from '../../components/InjectHTML/InjectHTML';
 
 const ActivityMain = () => {
-  const { currentModule, handleAnswers } = useContext(ActivityContext);
+  const { currentModule, handleAnswers, sessionData } = useContext(SessionContext);
 
-  const { sessionData } = useContext(SessionContext);
+
 
   // useEffect(() => {
   //   // touchSection();
