+++ src/pages/Activity/Activity.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
-import { ActivityContext } from '../../contexts/ActivityContext';
 import {UserContext} from '../../contexts/UserContext'
+import {SessionContext} from '../../contexts/SessionContext'
 import * as U from '../../components/TuxComponents/UniversalComponents';
 import ActivityHeader from '../../components/ActivityHeader/ActivityHeader';
 import ActivityBody from '../../components/ActivityBody/ActivityBody';
@@ -8,7 +8,7 @@ import SideBarNav from '../../components/SideBarNav/SideBarNav';
 import Overview from '../../components/Overview/Overview';
 
 const Activity = () => {
-  const { started } = useContext(ActivityContext);
+  const { started } = useContext(SessionContext);
   const {user} = useContext(UserContext)
 
   return (
